disabled krono in the char flags by default charsheet work

+++ trunk/source/common/Crypto.cpp
@@ -30,7 +30,7 @@ uint64_t Crypto::RSADecrypt(unsigned char* text, uint16_t size) {
     if (bLittleEndian) {
         ret = ntohl(*reinterpret_cast<uint32_t*>(text));
         ret <<= 32;
-        ret |= static_cast<uint64_t>(ntohl(*reinterpret_cast<uint32_t*>(text + 4)));
+        ret |= ntohl(*reinterpret_cast<uint32_t*>(text + 4));
     }
     else {
         ret = *reinterpret_cast<uint64_t*>(text);
